Refactored middleware
* Removed ctxroot middleware. Is now a internal helper function used in Proxy handler
* Renamed JWKS middlware to OIDC
* Moved JWKS logic to OIDC middleware
* Decoupled proxy from middlewares
