--- pkg/middleware/middleware_test.go
@@ -1,1 +1,0 @@
-package middleware
