--- pkg/proxy/middleware.go
@@ -1,329 +1,0 @@
-package proxy
-
-import (
-	"bufio"
-	"context"
-	"crypto/rsa"
-	"encoding/base64"
-	"fmt"
-	"github.com/SermoDigital/jose/crypto"
-	"github.com/SermoDigital/jose/jws"
-	"github.com/opentracing/opentracing-go"
-	"github.com/prometheus/client_golang/prometheus"
-	"gopkg.in/square/go-jose.v2/jwt"
-	"log"
-	"math/big"
-	"net"
-	"net/http"
-	"net/url"
-	"strings"
-	"time"
-)
-
-type ctxKey string
-
-var (
-	contextKey = ctxKey("Context")
-	subjectKey = ctxKey("Subject")
-)
-
-var frontendGauge = prometheus.NewGauge(prometheus.GaugeOpts{
-	Name: "multikube_frontend_live_requests",
-	Help: "A gauge of live requests currently in flight from clients",
-})
-
-var frontendCounter = prometheus.NewCounterVec(
-	prometheus.CounterOpts{
-		Name: "multikube_frontend_requests_total",
-		Help: "A counter for requests from clients",
-	},
-	[]string{"code", "method"},
-)
-
-var frontendHistogram = prometheus.NewHistogramVec(
-	prometheus.HistogramOpts{
-		Name:    "multikube_frontend_request_duration_seconds",
-		Help:    "A histogram of request latencies from clients.",
-		Buckets: prometheus.DefBuckets,
-	},
-	[]string{"handler", "method"},
-)
-
-// responseSize has no labels, making it a zero-dimensional
-// ObserverVec.
-var responseSize = prometheus.NewHistogramVec(
-	prometheus.HistogramOpts{
-		Name:    "multikube_frontend_response_size_bytes",
-		Help:    "A histogram of response sizes for client requests",
-		Buckets: []float64{200, 500, 900, 1500},
-	},
-	[]string{},
-)
-
-// Middleware represents a multikube middleware
-type Middleware func(*Proxy, http.Handler) http.Handler
-
-// responseWriter implements http.ResponseWriter and adds status code and response length bytes
-// so that WithLogging middleware can log response status codes
-type responseWriter struct {
-	http.ResponseWriter
-	status int
-	length int
-}
-
-// responseError satisfies the error interface
-type responseError struct {
-	Status int      `json:"status"`
-	Errs   []string `json:"errors"`
-}
-
-func init() {
-	prometheus.MustRegister(frontendGauge, frontendCounter, frontendHistogram, responseSize)
-}
-
-// getTokenFromRequest returns a []byte representation of JWT from an HTTP Authorization Bearer header
-func getTokenFromRequest(req *http.Request) []byte {
-	if ah := req.Header.Get("Authorization"); len(ah) > 7 && strings.EqualFold(ah[0:7], "BEARER ") {
-		return []byte(ah[7:])
-	}
-	return nil
-}
-
-// WriteHeader sends and sets an HTTP response header with the provided
-// status code. Implements the http.ResponseWriter interface
-func (r *responseWriter) WriteHeader(statusCode int) {
-	r.status = statusCode
-	r.ResponseWriter.WriteHeader(statusCode)
-}
-
-// Write implements the http.ResponseWriter interface
-func (r *responseWriter) Write(b []byte) (int, error) {
-	if r.status == 0 {
-		r.status = 200
-	}
-	n, err := r.ResponseWriter.Write(b)
-	r.length += n
-	return n, err
-}
-
-// Hijack implements the http.Hijacker interface
-func (r *responseWriter) Hijack() (net.Conn, *bufio.ReadWriter, error) {
-	if r.length < 0 {
-		r.length = 0
-	}
-	return r.ResponseWriter.(http.Hijacker).Hijack()
-}
-
-// WithEmpty is an empty handler that does nothing
-func WithEmpty(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		next.ServeHTTP(w, r)
-	})
-}
-
-// WithTracing is a middleware that starts a new span and populates the context
-func WithTracing(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		span := opentracing.GlobalTracer().StartSpan("hello")
-		ctx := opentracing.ContextWithSpan(r.Context(), span)
-		defer span.Finish()
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
-
-// WithLogging applies access log style logging to the HTTP server
-func WithLogging(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-		lrw := &responseWriter{ResponseWriter: w}
-		next.ServeHTTP(lrw, r)
-		var isResCached bool
-		if lrw.Header().Get("Multikube-Cache-Age") != "" {
-			isResCached = true
-		}
-		duration := time.Now().Sub(start)
-		log.Printf("%s %s %s %s %s %d %d %s %t", r.Method, r.URL.Path, r.URL.RawQuery, r.RemoteAddr, r.Proto, lrw.status, lrw.length, duration.String(), isResCached)
-	})
-}
-
-// WithJWT is a middleware that parses a JWT token from the requests and propagates
-// the request context with a claim value.
-func WithJWT(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		// Get the JWT from the request
-		t, err := jws.ParseJWTFromRequest(r)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		// Check if request has empty credentials
-		if t == nil {
-			http.Error(w, "No valid access token", http.StatusUnauthorized)
-			return
-		}
-
-		// Set context
-		username, ok := t.Claims().Get(c.OIDCUsernameClaim).(string)
-		if !ok {
-			username = ""
-		}
-
-		ctx := context.WithValue(r.Context(), subjectKey, username)
-
-		next.ServeHTTP(w, r.WithContext(ctx))
-
-	})
-}
-
-// WithRS256Validation is a middleware that validates a JWT token in the http request using RS256 signing method.
-// It will do so using a rsa public key provided in Config
-func WithRS256Validation(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		t, err := jws.ParseJWTFromRequest(r)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		if t == nil {
-			http.Error(w, "No token in request", http.StatusUnauthorized)
-			return
-		}
-
-		err = t.Validate(c.RS256PublicKey, crypto.SigningMethodRS256)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		// Set context
-		username, ok := t.Claims().Get(c.OIDCUsernameClaim).(string)
-		if !ok {
-			username = ""
-		}
-
-		ctx := context.WithValue(r.Context(), subjectKey, username)
-
-		next.ServeHTTP(w, r.WithContext(ctx))
-
-	})
-}
-
-// WithJWKValidation is a middleware that validates a JWT token in the http request using RS256 signing method.
-// It will do so using a JWK (Json Web Key) provided in c
-func WithJWKValidation(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		t, err := jws.ParseJWTFromRequest(r)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		raw := string(getTokenFromRequest(r))
-		tok, err := jwt.ParseSigned(raw)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		// Try to find a JWK using the kid
-		kid := tok.Headers[0].KeyID
-		jwk := c.JWKS.Find(kid)
-		if jwk == nil {
-			http.Error(w, "key id invalid", http.StatusUnauthorized)
-			return
-		}
-		if jwk.Kty != "RSA" {
-			http.Error(w, fmt.Sprintf("Invalid key type. Expected 'RSA' got '%s'", jwk.Kty), http.StatusUnauthorized)
-			return
-		}
-
-		// decode the base64 bytes for n
-		nb, err := base64.RawURLEncoding.DecodeString(jwk.N)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		// Check if E is big-endian int
-		if jwk.E != "AQAB" && jwk.E != "AAEAAQ" {
-			http.Error(w, fmt.Sprintf("Expected E to be one of 'AQAB' and 'AAEAAQ' but got '%s'", jwk.E), http.StatusUnauthorized)
-			return
-		}
-
-		pk := &rsa.PublicKey{
-			N: new(big.Int).SetBytes(nb),
-			E: 65537,
-		}
-
-		err = t.Validate(pk, crypto.SigningMethodRS256)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusUnauthorized)
-			return
-		}
-
-		username, ok := t.Claims().Get(c.OIDCUsernameClaim).(string)
-		if !ok {
-			username = ""
-		}
-
-		ctx := context.WithValue(r.Context(), subjectKey, username)
-
-		next.ServeHTTP(w, r.WithContext(ctx))
-
-	})
-}
-
-// WithHeader is a middleware that reads the value of the HTTP header "Multikube-Context"
-// in the request and, if found, sets it's value in the request context.
-func WithHeader(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		req := r
-		header := r.Header.Get("Multikube-Context")
-		if header != "" {
-			ctx := context.WithValue(r.Context(), contextKey, header)
-			req = r.WithContext(ctx)
-		}
-		next.ServeHTTP(w, req)
-	})
-}
-
-// WithCtxRoot is a middleware that reads the url path params in the request and
-// tries to determine which kubeconfig context to use for upstream api server requests.
-// If a context is found in the URL path params, the request-context is populated with the value
-// so that other handlers and middlewares may use the information
-func WithCtxRoot(c *Proxy, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		req := r
-		c, rem := getCtxFromURL(r.URL)
-		if c != "" {
-			ctx := context.WithValue(r.Context(), contextKey, c)
-			req = r.WithContext(ctx)
-			if rem != "" {
-				req.URL.Path = rem
-			}
-		}
-		next.ServeHTTP(w, req)
-	})
-}
-
-// getCtxFromURL reads path params from u and returns the kubeconfig context
-// as well as the path params used for upstream communication
-func getCtxFromURL(u *url.URL) (string, string) {
-	val := ""
-	rem := []string{}
-	if vals := strings.Split(u.Path, "/"); len(vals) > 1 {
-		val = vals[1]
-		rem = vals[2:]
-	}
-	return val, fmt.Sprintf("/%s", strings.Join(rem, "/"))
-}
